py_Crawler
This code can crab the data about Chinses stocks concerning about the social responsibilities. Because the data are not included in the source code, I try another way instead of the convenient way of crabbing form-format webs, coz it wont work. I find ports where store the data I want and use them as the urls. The other problem this code solved is that, this web's url wont change, however, using the method of ports, it wont bother us naturally.